Rename ReplHelper -> TermOutputHelper (#126)
